src/requirements.txt
pytorch>=1.0.1
ignite==0.2.1
h5py==2.7.1
